fix: Wrong Month and Temperature Units + index
Updated the month output as I didn't account for the month being
one off due to it being zero based. Updated the code to include
written month and day of the week instead of number dates only

Updated the temperature units to display imperial rather than Kelvin
and Celsius

Updated speech in index.html to instruct the user to type a 5 digit
code. May revert but unsure.